pyproject.toml
[project]
name = "securatron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsonschema>=4.26.0",
    "python-ulid>=3.1.0",
    "pyyaml>=6.0.3",
]
